// src/generated/models/Cr720_wfrdmanufactureaddressesesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr720_wfrdmanufactureaddressesesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr720_wfrdmanufactureaddressesesstatecode = keyof typeof Cr720_wfrdmanufactureaddressesesstatecode;
export const Cr720_wfrdmanufactureaddressesesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr720_wfrdmanufactureaddressesesstatuscode = keyof typeof Cr720_wfrdmanufactureaddressesesstatuscode;

export interface Cr720_wfrdmanufactureaddressesesBase {
  cr720_choice?: string;
  cr720_locationcep?: string;
  cr720_locationcidade?: string;
  cr720_locationcoordenadas?: string;
  cr720_locationestado?: string;
  cr720_locationnome?: string;
  cr720_locationpasregio?: string;
  cr720_locationrua?: string;
  cr720_name?: string;
  cr720_wfrdmanufactureaddressesid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Cr720_wfrdmanufactureaddressesesstatecode;
  statuscode?: Cr720_wfrdmanufactureaddressesesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr720_wfrdmanufactureaddresseses extends Cr720_wfrdmanufactureaddressesesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}